requirements/base.txt
confluent-kafka==2.3.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
prometheus-client==0.19.0
python-json-logger==2.0.7
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
pytz==2023.3
typing-extensions==4.8.0
asyncio-contextmanager==1.0.1
tenacity==8.2.3